Add worker thread to DnsProxyListener
Improve DnsProxy performance by adding a worker
thread that always runs.

DnsProxyWorker owns a thread and a queue. New
getaddrinfo requests and new gethostbyaddr requests
are put in the queue. When a new request is put in
queue DnsProxyWorker's thread is signaled that a
new job has been added. DnsProxyWorker's thread
gets the first job in the queue and executes its
handler.

With the patch it takes approx. 3 ms to execute
InetAddress.getByName("www.google.com") without the
patch it takes approx. 13 ms. With the java dns
cache disabled

Change-Id: I9712be09fbaec09fc80a5d13024bb06deabab6eb