Add worker threads to DnsProxyListener
Improve DnsProxy performance by adding a
worker threads that always runs.

DnsProxyWorkerPool holds a queue of DnsProxyJobs
and a number of worker threads. The getaddrinfo
and gethostbyaddr request are added as DnsProxyJobs
to the queue. The worker threads are reading from the
the queue and execute the request associated with the
job.

Max number of worker threads is 10.

With the patch it takes approx. 3 ms to execute
InetAddress.getByName("www.google.com") without the
patch it takes approx. 13 ms. With the java dns
cache disabled

Change-Id: I9712be09fbaec09fc80a5d13024bb06deabab6eb