Add a pool of threads that handles dns requests.
Improve DnsProxyListener performance by adding
a pool of threads that handles the request instead of
spawning a new thread for each request.

DnsProxyWorkerPool holds a queue of DnsProxyJobs
and a number of worker threads. The getaddrinfo
and gethostbyaddr request are added as DnsProxyJobs
to the queue. The worker threads are reading from the
the queue and execute the request associated with the
job.

This patch reduce the time a request spends in the
DnsProcyListener with about 50 percent.

Change-Id: I0ae951ef5c92741e765dafad7fbf5937f1eab9ed